move join/split cells out of coreActions.
